requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-multipart
python-dotenv
PyMuPDF==1.24.10
python-docx==1.2.1
Pillow==11.1.0
pytesseract==0.3.13
spacy==3.8.1
yake==0.7.3
textstat==0.7.13
langdetect==1.0.9
groq==0.14.2
requests
